Case insensitive json parsing
